added preview video implementation added random color picking fixed zone lines mouvement bug updated pom java version

+++ src/main/java/model/VideoType.java
@@ -11,7 +11,7 @@ public class VideoType extends VisualType {
 	public boolean muted;
 	public float clip_start_time;
 	public float clip_end_time;
-	public float duration;
+	public double duration;
 	
 	public VideoType() {
 	}
